ENH: class-skeleton: public parent <=> noncopyable
The snippet routine recognizes any public parent as a class that induces
non-copyability.

This is a simplification (as possibly all parents, and members) shall be
analyzed to be sure.
Let's just suppose the end-user knows that classes from public hierarchies
should be non copyable, and that he had used the base-class snippet to
generate its base classes, and never he will inherit publicly from a class not
designed to be inherited.

See CppCoreGuidelines C.67